Adds `Queryable` interface.
This provides an interface for an object used *only* for querying. This is step 1 in decoupling `comp.live` into foundational primitives which can be composed together easily. The implementation is mostly a copy of the existing `ComponentRef.prototype.query` and `ComponentRef.prototype.queryAll` functions.

I opted for an interface with separate implementation functions instead of a class. I initially went with a class approach but decided against it. This is because `Queryable` is inteded to be reused for a couple elements which query and return themselves instead of `Queryable` directly. This circular dependency led to a lot of trouble in the `Queryable` type and made overridding these methods much more complicated and inelegant. Instead an interface should be sufficient for typing needs. Separate implementation functions means they can slightly diverge from the contract of the interface (they don't return a `Queryable` or a real `Array` for instance) so subclasses of `Queryable` can more effectively reuse them without unnecessary allocations and transformations which would harm performance.