Adds `Dehydrated`.
This is step 2. on the path of decoupling `comp.live` into its fundamental primitives in a composable fashion. `Dehydrated` represents an element which has already been queried and found, but not yet hydrated. It deliberately obfuscates the element it represents, making it impossible to accidentally depend on methods or features of the underlying custom element which cannot be relied upon because the element hasn't been hydrated yet. For now, the only way to access the element is with `unvalidatedElement`, which is the escape hatch method which returns the element without any additional validation. However future methods will assert that the element doesn't need to be hydrated or actually hydrate it directly, making this much safer.

Beyond that, it implements `Queryable` so a dehydrated element can still be queried for its children. Ths implementation overrides the type slightly to return more `Dehydrated` elements, which should maximize the utility of this class. it makes sense that dehydrated elements can be queried and return other dehydrated elements.

One minor challenge is that it may not make all that much sense to query dehydrated elements. For example, an element could be prerendered with a `<span>` tag, but then at hydration time it could replace that with a completely different `<div>` tag. It would probably make more sense to say that the DOM structure of an element is private until *after* hydration. This way a parent element can't grab references to internals of a child element prior to hydrating it. However this is not easily solvable for a few reasons:
1. The nature of light DOM means that element internals and content projection into a child element are indistinguishable. We have no way of knowing whether a descendant `<span>` tag is actually owned by its direct parent, or an ancestor which projected the value into its child.
2. There may be legitimate use cases for interacting with the internal DOM of a child element prior to hydration.
3. If a component really cares about protecting its DOM this way, shadow DOM would be a great choice (once we eventually support shadow DOM).
4. Trying to prevent this would likely be more trouble than it's worth and just make things overall harder and more complicated.