Adds `ElementAccessor`.
This is step 3 on the path of decoupling `comp.live` into foundational, composable primitives. This is intended as an `Element` wrapper similar to `ElementRef`. For now it just supports querying, but the idea is to make it easy to read and enhance the underlying `Element` object by working with serializers and likely replacing `ElementRef` in the future.