Adds `bind` signal accessor.
This is step 4 on the plan to decouple `comp.live` into foundational, composable primitives. `bind` is now a separate function which takes an `ElementAccessor` as an input. I'm not 100% sold that this is the right design as I think it confuses order of operations by breaking chaining from a `Dehydrated` or `ElementAccessor`. However, this also allows `bind` to be tree-shaken if not used, meaning signals can potentially be dropped from the bundle entirely if they aren't needed. I'm going to at least give this DX try before overengineering it to a more complex solution.

This is mostly a port of the exiting `ComponentRef.prototype.bind`. Notably, this is *not* a direct copy and does include some functionality differences:
1.  `ComponentRef.prototype.bind` supported querying its children. This is handled by `Queryable.prototype.query` and does not need to be included here.
2.  `ComponentRef.prototype.bind` supported implicit serializers. A signal which returns a `string` would automatically use the `String` serializer. I decided not to keep this feature as I'm not convinced its a great idea and it is one more way in which primitive serializers are un-tree shakable.