Adds `live` signal accessor.
This is a standalone function just like `bind` but creating a `WriteableSignal` value which is reflected back into DOM when it is mutated. The implementation and tests are basically just copies of the existing `ComponentRef.prototype.live` with only minor changes such as removing query support since users should call `Queryable.prototype.query` instead.