Adds `isHydrated` and `hydrate` utilities.
These should provide reusable utilites for common problems and make it easier to identify the purpose of their use by the simple fact of having named them. Each utility does have a number of unique edge cases to handle, so doing this in one place makes it less likely to get those wrong in the future.

The main objective of each is to manipulate the `defer-hydration` attribute. The most interesting design used here is taking `elementClass` as an input to `hydrate`. This is not strictly necessary as all we do is an `instanceof` check. However it is useful to ensure the dependency is sorted first by any bundler. Since `customElements.define` is typically co-located with its class definition, that means it is highly likely that the class is defined prior to being hydrated, which is exactly what we want. This is technically not a hard requirement, `customElements.define` could be far away from the class definition and happen at a different time, however this is the best and most intuitive behavior we can create out of the box. `hydrate` does ensure that the element is defined and upgraded first, so even if this invariant is broken, the developer will still receive a useful error, however fixing it will heavily depend on their strategy for defining custom elements.

I initially tried to define a type like:

```typescript
type Dehydrated<El extends Element> = Element & {  __haHydratedType: El };
```

This type would very much *not* be a custom element, but could potentially be upgraded into one. The idea was that we could use this to define `hydrate` like so:

```typescript
declare function hydrate<El extends Element>(el: Dehydrated<El>, elementClass: typeof El): asserts el is El;
```

The advantage of this approach is that it means we can pass around `Dehydrated<MyComponent>` references safely. They are equivalent to `Element`, so there's no risk of accidentally using `MyComponent` APIs before it has hydrated. Then you can pass the element input `hydrate` and it will convert the type. It can also type check that the class you gave matches the type of the `Dehydrated` you gave it. This can kind of work (though not with that definition, it is more complicated than that).

However I ultimately decided not to do this for a few reasons:
1.  Having the `Dehydrated` type doesn't avoid needing the element class for hydration.
2.  Types get quite complicated here.
3.  Behavior gets confusing when you consider that `Dehydrated<Element>` could be upgraded with `MyComponent` or vice versa. Getting the errors to show correctly gets complicated.
4.  You can't assert a parameter to be a type which is not a subtype of the input. For example, you can't assert that a `string` parameter is actually a `number`. This means `__haHydratedType: El` is wrong, it needs to be optional (`__haHydratedType?: El`). That doesn't sound like a big deal, but it means that `Dehydrated<Element>` is assignable to `Element` which makes it very easy to accidentally escape the protections `Dehydrated` is supposed to prevent.
5.  I have a separate idea for `Dehydrated` as a class which wraps an element to hide its methods. That is coming up in a future commit. While technically a separate concept and the two could be used together, doing so would be confusing and there likely would not be a need for both.