Adds `Dehydrated.prototype.access` and `Dehydrated.prototype.hydrate`.
These provide safe accessors to assert or hydrate an element before using it and return the "hydrated by construction" `ElementAccessor` type.

I debated on the exact semantics of `access` for a while. I ultimately decided that the purpose of this function is to ensure that the element access is "safe" from hydration timing issues, so everything is focused around that. `.access(Element)` is prevented by the type system because the only reason to do this is to work around the required dependency, but we already have `dehydratedElement` for that use case. I also considered putting the hydration checks in `ElementAccessor.from` instead of `Dehydrated.prototype.access`. However I ended up feeling that `ElementAccessor` is just a convenient wrapper for an element which may or may not be hydrated. It's really `Dehydrated` in combination with its `access` and `hydrate` methods which provide strong guarantees that users don't accidentally depend on dehydrated elements.