Adds hydration composition example.
This uses `.access` and `.hydrate` to demonstrate their purpose. Currently HydroActive doesn't have a good way for providing public methods on the component object itself, so I used a native web component with manual `defer-hydration` integration to create one. HydroActive uses this element to illustrate that it naturally enforces that all observed elements are upgraded and hydrated prior to use.