Adds `shadow` to `Queryable` interface and `QueryRoot` implementation.
This provides an accessor for shadow roots and allows a DX like `host.shadow.query('.foo')`.

As much as I like shadow DOM and think it is a good default option, I am currently in the minority on that front. So instead of checking shadow DOM by default, I'm requiring a `.shadow` accessor. This roughly aligns with how native DOM APIs work (`el.querySelector` vs `el.shadowRoot!.querySelector`).

I debated for a while whether `QueryRoot.prototype.shadow` should return a `Queryable` or a `QueryRoot` (and the same question also applies to `Dehydrated` and `ElementAccessor`). The challenge here is that `Queryable.prototype.shadow` returns `Queryable` which queries other `Queryable` objects. But `Dehydrated` and `ElementAccessor` both need it to query `Dehydrated` elements. `QueryRoot` is compatible with both an `Element` and a `ShadowRoot`, but the same can't be said for `Dehydrated` or `ElementAccessor` which must wrap `Element` objects. As a result, `Dehydrated.prototype.shadow` can't return `Queryable` (because it needs to query `Dehydrated` elements), but also can't wrap it in another `Dehydrated` because you can't have a `Dehydrated<ShadowRoot>`. I briefly considered a `DehydratedRoot` which would wrap `Queryable` and return `Dehydrated` values, but this feels overly complex for the problem at hand. Instead, I had `QueryRoot` query `Dehydrated` elements and expose `QueryRoot` for both `{Dehydrated,ElementAccessor}.prototype.shadow`. This is a slightly leakier abstraction than it needs to be and also makes `QueryRoot` less independently useful (`QueryRoot.from(el).query('...')` somewhat confusingly returns a `Dehydrated`), but I think this is an acceptable balance of the trade off between semantic accuracy and simplicity / code size.

This approach only supports `open` shadow roots as we need a custom element to get access to `closed` shadow roots. A separate commit will look into adding support for `closed` HydroActive elements at least.