Adds `ComponentAccessor` to support closed shadow roots.
This is a subclass of `ElementAccessor` specifically for `HydroActiveComponent` elements. The reason for this is because that type specifically now automatically captures and exposed closed shadow roots via its `ElementInternals` object. By doing this, `ComponentAccessor.fromComponent(comp)` is able to automatically support closed shadow roots.

As a large comment describes, we need to retain the reference to `ElementInternals` instead of just the `ShadowRoot` because a new shadow root can be attached at any time, not just construction time, and when that happens the value will magically appear on `ElementInternals`. So we need to pass this through in the `getClosedShadowRoot` argument to ensure we check `ElementInternals` for any query.

I also went out of my way to store the `ElementInternals` on a `WeakMap` which is *not* public API so the closed shadow root is truly "closed". No one outside the component can get a reference to its `ElementInternals` or the shadow root. Putting this information on a property of the component (even if "private" like `HydroActiveComponent.prototype.__elementInternals`) would leak this closed shadow root. If the user went out of their way to close the shadow root, then it's probably important to keep the constraint that the root is truly inaccessible outside the component.