Adds `QueryRoot`.
This stands a base implementation of `Queryable` which can be reused and composed by other classes like `Dehydrated` and `ElementAccessor`. `QueryRoot` is distinct from `Queryable` because it allows classes to implement `Queryable` without necessarily buying into excess baggage of `QueryRoot` (such the `root` accessor). This keeps things slightly decoupled and a little more flexible.

For now, I'm omitting `.query` and `.queryAll` tests for now as it will soon be adopting the existing implementation from `queryable.ts`.