Updates `Dehydrated` and `ElementAccessor` to compose `QueryRoot`.
This is a more natural extension point than the loose `query` and `queryAll` functions. I opted to compose `QueryRoot` instead of extend it as `QueryRoot` includes symbols like `root` which expose the wrapped node. This goes directly against the design goals of `Dehydrated` in particular which exists to protect that element as much as possible unless going through known-safe accessors. `root` as a term also doesn't mean a whole lot for `Dehydrated` and `ElementAccessor` anyways.