Adds `out` to generic types to ensure they remain covariant and not contravariant.
All of these classes return types based on their roots, so they must be covariant to support that. However accidentally putting the root in a parameter position would make them contravariant as well (and become invariant as a result). This would make them very restrictive to work with. By adding `out` we're explicitly denoting the types to be covariant (and therefore not contravariant) so accidentally adding such a function would result in a clear error at definition rather than usage.

I made this mistake earlier and I think it's good to enforce this constraint to make the contract of these types a little bit clearer.