Bump TypeScript to v5.5.4 (current latest).
This required one minor change to `QueriedElement` where a `null` seemed to lead to a `Dehydrated<null>` type in `query-root.ts` which is not allowed. I'm not entirely sure why that `null` doesn't get properly filtered out, but this seems like an easy fix and doesn't make the type meaningfully less accurate, since `:scope` is always `null` when used on a `shadowRoot` and `QueriedElement` results are implicitly `null` anyways given that we can never guarantee an element which actually be found at runtime.