Adds `ReactiveRoot`.
This interface describes an object which can schedule effects against a `Connectable` lifecycle. Practically speaking, this allows effects to be automatically bound to a component being connected to / disconnected from the document. The implementation and tests are mostly copied over from `ComponentRef`.

Having this as a standalone implementation allows this to be more effectively composed throughout the system.