Adds `defineBaseComponent`.
This is a non-signal version of `defineSignalComponent`. The reduced dependencies means the component is lighter weight at the cost of being natively incompatible with signal APIs. Specifically, the provided `host` is a `ComponentAccessor`, not a `SignalComponentAccessor`, so it does not support `.effect` and cannot be used as a `ReactiveRoot` for signal APIs.

I considered naming this `defineComponent` but decided against it because I would prefer for `defineSignalComponent` to be true "default" option. `defineBaseComponent` should only be used in exceptional cases where the component is simple enough to not need the weight of a signals implementation. Most use cases won't benefit from it and I would rather most devs think about `defineSignalComponent` as the primary entry point into HydroActive.