Updates demo components to use `defineBaseComponent` where appropriate.
Any components which don't use signals don't need `defineSignalComponent`. I was a little torn on this just because I want developers to think of `defineSignalComponent` as the "default" option and would rather not draw too much attention to `defineBaseComponent`. That said, these examples should be as minimal as possible, so eliminating signals from their dependencies seems like a good move and it doesn't really make things less clear.

The main challenge is that someone might copy an example with `defineBaseComponent` and then copy a `.effect` call or a `live` binding from another component and then be confused when it doesn't work. That's probably more of a documentation issue than one with the examples, so I'm inclined to used `defineBaseComponent` where possible for now.