Sets `sideEffects: false` in `package.json`.
This option is read by bundlers like esbuild and Webpack to mean that the entire package is to be treated as if it contains no top-level side effects. This allows bundlers to tree shake unused code in the top-level scope. Without this, code like `const defaultScheduler = MacrotaskScheduler.from();` needs to be retained, even when `defaultScheduler` is never used purely because the bundler cannot prove it to not have meaningful side effects. `sideEffects: false` allows the bundler to eliminate this code.

https://webpack.js.org/guides/tree-shaking/#mark-the-file-as-side-effect-free