Applies a component definition returned by the `hydrate` function of `defineBaseComponent` / `defineSignalComponent`.
This allows a component to export public properties on the resulting component.

I'm still torn about this because it creates a new instance of each property for each component *instances*. Normally you would want a single prototype containing all the functions shared amongst all instances (a class prototype), as this would be easier to optimize for the engine. I don't see a better solution right now for the functional DX, but I'll have to keep playing with the class authoring format to see if an alternative design is feasible.