Adds `StabilityTracker`.
This is mostly a copy of the existing stability implementation in `Scheduler`, this new version will obsolete the one in `Scheduler`.

The goal here is to decouple stability tracking from scheduling, allowing multiple `Scheduler` objects to be tracked as part of the same `.stable()` call. This allows one component to have a single `StabilityTracker` which tracks actions scheduled from many different `Scheduler` objects.