Supports synchronous scheduling in `effect`.
This broke for all signal changes after the first when using a synchronous `Scheduler` because `cancelNextCall = undefined` in the `.schedule` callback was getting invoked *before* the assignment to `cancelNextCall = scheduler.schedule(() => /* ... */)`, meaning the cancel function always exists. Since we used `cancelNextCall` to determine whether an operation was scheduled, `effect` incorrectly thought the action was still scheduled when it had actually already executed.

This is fixed by using a distinct `scheduled` boolean which looks a bit redundant, but is necessary for the synchronous `Scheduler` case. Added a regression test to be more clear about this requirement.