Drops `define` from component creation function names.
These functions no longer "define" the component because they do not call `customElements.define` automatically. I debated the names for a while but the best I could come up with was to revert to the previous name of `component` for signal components and introduce `baseComponent` for base components.

I considered `declare*Component`, but I think it's too easy to confuse `define` and `declare`. Other verbs like `create` or `make` imply creating a component instances rather than a component class.